Remove refresh button from Glitch site
The current implementation just makes a regular POST request to the
refresh endpoint, but that's not actually how the browser does refresh
requests. We could probably implement forced refreshes by deleting some
cookies and reloading the page, but that can be done manually or with
short-expiring bound cookies. We'll implement that feature if we need
it.