Add challenges to Glitch server (#96)
* Validate JWT in session registration and store keys

We currently don't extract the key from the JWT or check that the JWT
was signed correctly. We should do that so we can actually issue
challenges in the future.

* Issue challenges at refresh time

We currently don't validate the session is from the same device at
refresh time. This CL adds a round of challenge verification.

---------

Co-authored-by: Daniel Rubery <drubery@chromium.org>