Remove Sec- prefixes
This is meaningless on the response headers, and on the
Sec-Session-Response, the server needs to carefully validate that the
expected key is being used anyway. It does seem kind of nice if the
server can trust that the Sec-Session-Id is truly coming from the site
instead of an XHR, so that it can look up the right expected public key
for the Sec-Session-Response, so we leave the Sec- prefix on that one
header. Following the discussion on
https://github.com/w3c/webappsec-dbsc/issues/59, we make the shared
prefix Secure-Session- instead.