Only allow refresh of same session
This PR makes a few changes to how we do refreshes. We now only allow
refreshes to change the params for the session being refreshed. We also
allow refresh endpoints to not serve any config. This PR is also a
little stricter with terminating the session on invalid refresh configs.