Update redirect handling for DBSC requests
The current description doesn't really align with the Fetch redirect
handling. While it's nice to ensure that sites don't expose themselves
to significant risk, we already require that the refresh URL be HTTPS,
so sites need to do significant work to take on this risk. Simpler Fetch
integration is valued over protecting sites from themselves.