Change some field defaults (#182)
* Defaults for session scope entries

Sites shouldn't have to specify both domain and path in every rule. By
default, these should always match.

* Make include_site mandatory

Most site owners are going to have to think through the scope of their
session anyway, so this is not a significant burden. With a mandatory
field, we can consider combining it into `session_inclusion_rules` in a
backwards-compatible way.

---------

Co-authored-by: Daniel Rubery <drubery@chromium.org>