Improve key sharing section (#181)
* Replace Identity Provider with Session Provider

We don't want to make it seem like DBSC is deeply tied to login, given
that it is actually only tied to cookies. So we coin the term Session
Provider (SP) instead.

* Require RPs to declare their SPs

In order to prevent an RP from using multiple SPs, we need to define
links from RP to their one and only SP.

* Use permissions spec for key sharing

By treating this as a permission, browsers are free to choose the UI
that makes sense for them.

* Address comments

---------

Co-authored-by: Daniel Rubery <drubery@chromium.org>