Re-run challenge header processing after creating a session
To relieve site developers from having to understand the nuances of DBSC's
fetch integration, we want to allow the registration response to do both
registration and set a challenge.

This change is needed because normally, the challenge header is
processed before registration, and at that point there is no session to
cache the requested challenge onto.